# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CameraBased API",
    "version": "v1"
  },
  "paths": {
    "/Account/Get": {
      "get": {
        "tags": [
          "Account"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Account"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Account/GetByName/{name}": {
      "get": {
        "tags": [
          "Account"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Account"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
    "/Account/GetById/{id}": {
      "get": {
        "tags": [
          "Account"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Account/Create": {
      "post": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateAccountDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Account/Update/{id}": {
      "put": {
        "tags": [
          "Account"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AccountDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Account/Delete/{id}": {
      "delete": {
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
    "/SubAccount/GetByName/{name}": {
      "get": {
        "tags": [
          "SubAccount"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SubAccount"
              }
            }
          }
        }
      }
    },
    "/SubAccount/GetById/{id}": {
      "get": {
        "tags": [
          "SubAccount"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SubAccount"
            }
          }
        }
      }
    },
    "/SubAccount/GetByPhone/{id}": {
      "get": {
        "tags": [
          "SubAccount"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/SubAccount/Create": {
      "post": {
        "tags": [
          "SubAccount"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SubAccountDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SubAccount"
            }
          }
        }
      }
    },
    "/SubAccount/Update/{id}": {
      "put": {
        "tags": [
          "SubAccount"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SubAccountDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/SubAccount/Delete/{id}": {
      "delete": {
        "tags": [
          "SubAccount"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "definitions": {
    "Account": {
      "required": [
        "accounName",
        "password"
      ],
      "type": "object",
      "properties": {
        "accountID": {
          "format": "int32",
          "type": "integer"
        },
        "accounName": {
          "maxLength": 100,
          "type": "string"
        },
        "password": {
          "maxLength": 100,
          "type": "string"
        },
        "accountEmail": {
          "maxLength": 200,
          "type": "string"
        },
        "phone": {
          "maxLength": 20,
          "type": "string"
        },
        "fullName": {
          "maxLength": 100,
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "roleID": {
          "format": "int32",
          "type": "integer"
        },
        "role": {
          "$ref": "#/definitions/Role"
        },
        "notifiHistories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NotifiHistory"
          }
        },
        "subAccounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubAccount"
          }
        },
        "refreshToken": {
          "type": "string"
        },
        "tokenCreated": {
          "format": "date-time",
          "type": "string"
        },
        "tokenExpires": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "AccountDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "accountEmail": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "roleID": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "Carlocator": {
      "type": "object",
      "properties": {
        "carLocatorID": {
          "format": "int32",
          "type": "integer"
        },
        "location": {
          "type": "string"
        },
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        },
        "carManagement": {
          "$ref": "#/definitions/CarManagement"
        }
      },
      "additionalProperties": false
    },
...
    "CarManagement": {
      "type": "object",
      "properties": {
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        },
        "carName": {
          "maxLength": 100,
          "type": "string"
        },
        "carColor": {
          "maxLength": 100,
          "type": "string"
        },
        "licensePlates": {
          "maxLength": 100,
          "type": "string"
        },
        "carBrand": {
          "maxLength": 100,
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "notifiHistories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NotifiHistory"
          }
        },
        "carlocators": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Carlocator"
          }
        }
      },
      "additionalProperties": false
    },
...
    "CreateAccountDTO": {
      "type": "object",
      "properties": {
        "accountName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "roleID": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "NotifiHistory": {
      "type": "object",
      "properties": {
        "historyID": {
          "format": "int32",
          "type": "integer"
        },
        "historyName": {
          "maxLength": 100,
          "type": "string"
        },
        "historyDate": {
          "format": "date-time",
          "type": "string"
        },
        "historyStatus": {
          "type": "boolean"
        },
        "accountID": {
          "format": "int32",
          "type": "integer"
        },
        "account": {
          "$ref": "#/definitions/Account"
        },
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        },
        "carManagement": {
          "$ref": "#/definitions/CarManagement"
        }
      },
      "additionalProperties": false
    },
    "Role": {
      "required": [
        "roleName"
      ],
      "type": "object",
      "properties": {
        "roleID": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "maxLength": 30,
          "type": "string"
        },
        "accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Account"
          }
        }
      },
      "additionalProperties": false
    },
    "SubAccount": {
      "type": "object",
      "properties": {
        "subAccountID": {
          "format": "int32",
          "type": "integer"
        },
        "subAccountName": {
          "type": "string"
        },
        "subAccountPhone": {
          "type": "string"
        },
        "accountID": {
          "format": "int32",
          "type": "integer"
        },
        "account": {
          "$ref": "#/definitions/Account"
        }
      },
      "additionalProperties": false
    },
    "SubAccountDTO": {
      "type": "object",
      "properties": {
        "subAccountName": {
          "type": "string"
        },
        "subAccountPhone": {
          "type": "string"
        },
        "accountID": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header uisng the Bearer scheme (\"bearer {token}\")"
    }
  }
}